feat: add unittest
