+++ test/vue/util.ts
@@ -1,6 +1,6 @@
 /* eslint-disable vue/one-component-per-file */
-import type { Component, InjectionKey, Ref } from 'vue'
-import { KeepAlive, createApp, createSSRApp, defineComponent, h, provide, ref } from 'vue'
+import type { App, Component } from 'vue'
+import { KeepAlive, createApp, createSSRApp, defineComponent, h, ref } from 'vue'
 import { renderToString } from '@vue/server-renderer'
 import { VueHeadMixin, createHead } from '@unhead/vue'
 import { renderSSRHead } from '@unhead/ssr'
@@ -39,9 +39,12 @@ export async function ssrRenderOptionsHead(input: any) {
 type InstanceType<V> = V extends { new (...arg: any[]): infer X } ? X : never
 type VM<V> = InstanceType<V> & { unmount(): void }
 
-export function mount<V extends Component>(Comp: V) {
+export function mount<V extends Component>(Comp: V, hook?: (p: { app: App }) => any) {
   const el = document.createElement('div')
   const app = createApp(Comp)
+  const uses = hook?.({ app })
+  if (uses)
+    Object.values(uses).forEach(value => app.use(value as any))
 
   const unmount = () => app.unmount()
   const comp = app.mount(el) as any as VM<V>
@@ -49,9 +52,7 @@ export function mount<V extends Component>(Comp: V) {
   return comp
 }
 
-export const Key: InjectionKey<Ref<number>> = Symbol('num')
-
-export function useKeepAliveSetup<V>(setup: () => V) {
+export function useKeepAliveSetup<V>(setup: () => V, hook?: (p: { app: App }) => any) {
   const Comp = defineComponent({
     setup,
     render() {
@@ -63,13 +64,12 @@ export function useKeepAliveSetup<V>(setup: () => V) {
     components: Comp,
     setup() {
       const visible = ref(true)
-      provide(Key, ref(1))
       return { visible }
     },
     render() {
       return h('div', [h(KeepAlive, [this.visible ? h(Comp, { ref: 'comp' }) : ''])])
     },
   })
-  const app = mount(Provider)
+  const app = mount(Provider, hook)
   return app
 }
